[tests] Run tests for event loop policy inside Pytester to avoid pollution of other tests with a custom event loop.
